Fix notification view for multiple notifications
When more notifications are shown that can fit on the screen, half of
the last notification was previously cut off margin has not been added
to a notification container class to fix this.

The notification / Announcement toggle has now been set to fixed to
allow the notifications to scroll below it so it is always in view.