Fix edge case where user has over 8 unread notifications
If a user has over 8 unread notifications the system will delete the
additional ones and never display them to the user, this commit ensures
all unread notifications are displayed and will also show an additional
8 read notifications to the user.

The additional 8 read notifications may be altered to fit the framework
better, this has just been left at this number it was before. These
read notifications should possibly be marked in some way in the
notification view to show they have already been read.

In a future commit adding a system to archive notifications instead of
delete them may be advisable.