Fix issue #79 - Trial period notification still subscription.
Add new column to notifications table to store the event the
notification is related to. This is used to track down a notification
at a later date which was generated for a specific event.

Update CreateTrialEndingNotification for single users and teams to
complete the event column on the notification.

Update CreateInvitationNotification for teams to complete the event
column on the notification.

Create RemoveTrialEndingNotification for single users and teams to
remove the Trial Ending Notification create at registration.