Fixed marking any notification as read.
Right now, if you pass any list of IDs in on the request you will mark them read. This means you could send a request to this endpoint and mark someone else's notifications as read.

This fix makes sure that you are only marking IDs owned by the authenticated user as read.