Update User notification relationship to hasMany
Currently, the User class uses the trait 'Notifiable', which uses a morphMany relationship via 'notifiable_id' and 'notifiable_type'. However, the notifications table does not take this into consideration. Also, there are several methods in other classes that include a 'user_id' parameter when creating notifications, which does not work with the morphMany relationship.
I figured that either Spark needed a complete overhaul of its Notification component to update to the Laravel 5.3 Notifications, or this quick fix would do the trick.
There seems to be only one place in the codebase where the 'notifications' method is called on a User instance. This is in the 'markAsRead' method of the NotificationController. When this method is called, it causes an exception to be thrown and the notifications don't get marked as read.